+++ R/spct.spikes.R
@@ -377,8 +377,10 @@ despike.generic_spct <-
                        y.var.name = y.var.name,
                        var.name = var.name,
                        ...)
-      return(rbindspct(mspct, idfactor = FALSE, attrs.simplify = TRUE))
+      return(rbindspct(mspct, idfactor = getIdFactor(x), attrs.simplify = TRUE))
     }
+
+    check_wl_stepsize(x)
 
     if (is.null(var.name)) {
       # find target variable
@@ -389,11 +391,7 @@ despike.generic_spct <-
     if (length(var.name) == 0L) {
       warning("No data columns found, skipping.")
     }
-    wl.step.range <- range(wl_stepsize())
-    if (wl.step.range[2] / wl.step.range[1] > 2.5) {
-      warning("Despike assumes consistent w.length steps! max step / min step = ",
-              wl.step.range[2] / wl.step.range[1])
-    }
+
     for (col.name in var.name) {
       if (!is.numeric(x[[col.name]])) {
         next()
@@ -442,8 +440,10 @@ despike.source_spct <-
                        na.rm = na.rm,
                        unit.out = unit.out,
                        ...)
-      return(rbindspct(mspct, idfactor = FALSE, attrs.simplify = TRUE))
+      return(rbindspct(mspct, idfactor = getIdFactor(x), attrs.simplify = TRUE))
     }
+
+    check_wl_stepsize(x)
 
     if (unit.out == "energy") {
       z <- q2e(x, action = "replace", byref = FALSE)
@@ -494,8 +494,10 @@ despike.response_spct <-
                        na.rm = na.rm,
                        unit.out = unit.out,
                        ...)
-      return(rbindspct(mspct, idfactor = FALSE, attrs.simplify = TRUE))
+      return(rbindspct(mspct, idfactor = getIdFactor(x), attrs.simplify = TRUE))
     }
+
+    check_wl_stepsize(x)
 
     if (unit.out == "energy") {
       z <- q2e(x, action = "replace", byref = FALSE)
@@ -548,8 +550,10 @@ despike.filter_spct <-
                        na.rm = na.rm,
                        filter.qty = filter.qty,
                        ...)
-      return(rbindspct(mspct, idfactor = FALSE, attrs.simplify = TRUE))
+      return(rbindspct(mspct, idfactor = getIdFactor(x), attrs.simplify = TRUE))
     }
+
+    check_wl_stepsize(x)
 
     if (filter.qty == "transmittance") {
       z <- A2T(x, action = "replace", byref = FALSE)
@@ -599,8 +603,10 @@ despike.reflector_spct <- function(x,
                      method = method,
                      na.rm = na.rm,
                      ...)
-    return(rbindspct(mspct, idfactor = FALSE, attrs.simplify = TRUE))
+    return(rbindspct(mspct, idfactor = getIdFactor(x), attrs.simplify = TRUE))
   }
+
+  check_wl_stepsize(x)
 
   col.name <- "Rfr"
   x[[col.name]] <- despike(x[[col.name]],
@@ -641,8 +647,10 @@ despike.solute_spct <-
                        method = method,
                        na.rm = na.rm,
                        ...)
-      return(rbindspct(mspct, idfactor = FALSE, attrs.simplify = TRUE))
+      return(rbindspct(mspct, idfactor = getIdFactor(x), attrs.simplify = TRUE))
     }
+
+    check_wl_stepsize(x)
 
     cols <- intersect(c("K.mole", "K.mass"), names(x))
     if (length(cols) == 1) {
@@ -687,8 +695,10 @@ despike.cps_spct <- function(x,
                      method = method,
                      na.rm = na.rm,
                      ...)
-    return(rbindspct(mspct, idfactor = FALSE, attrs.simplify = TRUE))
+    return(rbindspct(mspct, idfactor = getIdFactor(x), attrs.simplify = TRUE))
   }
+
+  check_wl_stepsize(x)
 
   var.name <- grep("cps", colnames(x), value = TRUE)
   for (col.name in var.name) {
@@ -730,8 +740,10 @@ despike.raw_spct <- function(x,
                      method = method,
                      na.rm = na.rm,
                      ...)
-    return(rbindspct(mspct, idfactor = FALSE, attrs.simplify = TRUE))
+    return(rbindspct(mspct, idfactor = getIdFactor(x), attrs.simplify = TRUE))
   }
+
+  check_wl_stepsize(x)
 
   var.name <- grep("counts", colnames(x), value = TRUE)
   for (col.name in var.name) {
@@ -1102,8 +1114,10 @@ spikes.generic_spct <-
                       na.rm = na.rm,
                       var.name = var.name,
                       ...)
-      return(rbindspct(mspct, idfactor = FALSE, attrs.simplify = TRUE))
+      return(rbindspct(mspct, idfactor = getIdFactor(x), attrs.simplify = TRUE))
     }
+
+    check_wl_stepsize(x)
 
     if (is.null(var.name)) {
       # find target variable
@@ -1152,8 +1166,10 @@ spikes.source_spct <-
                       na.rm = na.rm,
                       unit.out = unit.out,
                       ...)
-      return(rbindspct(mspct, idfactor = FALSE, attrs.simplify = TRUE))
+      return(rbindspct(mspct, idfactor = getIdFactor(x), attrs.simplify = TRUE))
     }
+
+    check_wl_stepsize(x)
 
     if (unit.out == "energy") {
       z <- q2e(x, "replace", FALSE)
@@ -1198,8 +1214,10 @@ spikes.response_spct <-
                       na.rm = na.rm,
                       unit.out = unit.out,
                       ...)
-      return(rbindspct(mspct, idfactor = FALSE, attrs.simplify = TRUE))
+      return(rbindspct(mspct, idfactor = getIdFactor(x), attrs.simplify = TRUE))
     }
+
+    check_wl_stepsize(x)
 
     if (unit.out == "energy") {
       z <- q2e(x, "replace", FALSE)
@@ -1246,8 +1264,10 @@ spikes.filter_spct <-
                       na.rm = na.rm,
                       filter.qty = filter.qty,
                       ...)
-      return(rbindspct(mspct, idfactor = FALSE, attrs.simplify = TRUE))
+      return(rbindspct(mspct, idfactor = getIdFactor(x), attrs.simplify = TRUE))
     }
+
+    check_wl_stepsize(x)
 
     if (filter.qty == "transmittance") {
       z <- A2T(x, "replace", FALSE)
@@ -1288,8 +1308,10 @@ spikes.reflector_spct <- function(x,
                     max.spike.width = max.spike.width,
                     na.rm = na.rm,
                     ...)
-    return(rbindspct(mspct, idfactor = FALSE, attrs.simplify = TRUE))
+    return(rbindspct(mspct, idfactor = getIdFactor(x), attrs.simplify = TRUE))
   }
+
+  check_wl_stepsize(x)
 
   col.name <- "Rfr"
   spikes.idx <-
@@ -1323,8 +1345,10 @@ spikes.solute_spct <-
                       max.spike.width = max.spike.width,
                       na.rm = na.rm,
                       ...)
-      return(rbindspct(mspct, idfactor = FALSE, attrs.simplify = TRUE))
+      return(rbindspct(mspct, idfactor = getIdFactor(x), attrs.simplify = TRUE))
     }
+
+    check_wl_stepsize(x)
 
     cols <- intersect(c("K.mole", "K.mass"), names(x))
     if (length(cols) == 1) {
@@ -1365,8 +1389,10 @@ spikes.cps_spct <- function(x,
                     na.rm = na.rm,
                     var.name = var.name,
                     ...)
-    return(rbindspct(mspct, idfactor = FALSE, attrs.simplify = TRUE))
+    return(rbindspct(mspct, idfactor = getIdFactor(x), attrs.simplify = TRUE))
   }
+
+  check_wl_stepsize(x)
 
   spikes.idx <-
     which(find_spikes(x[[var.name]],
@@ -1400,8 +1426,10 @@ spikes.raw_spct <- function(x,
                     na.rm = na.rm,
                     var.name = var.name,
                     ...)
-    return(rbindspct(mspct, idfactor = FALSE, attrs.simplify = TRUE))
+    return(rbindspct(mspct, idfactor = getIdFactor(x), attrs.simplify = TRUE))
   }
+
+  check_wl_stepsize(x)
 
   spikes.idx <-
     which(find_spikes(x[[var.name]],